backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==23.5
psycopg2-binary==2.9.9
celery==5.3.6
redis==5.0.1
cryptography==42.0.2
channels==4.0.0
channels-redis==4.2.0
daphne==4.1.0
python-dotenv==1.0.0
kubernetes==29.0.0
paramiko==3.4.0
ansible-runner==2.3.4
ansible==9.1.0
PyYAML==6.0.1
